Create a graph factory function to represent chessboard
There are too few commits for this project. Uses previously
implemented LinkedList and HashMap factory functions to help represent
the Graph as an adjacency list. The vertices of the graph are kept in
a Hash Map and each respective vertex' adjacent neighbors are kept in
a Linked List variable inside the Vertex factory function. A shortest
path function is implemented in the graph factory function using a
breadth-first search approach in order to serve as the basis for the
shortest amount of Knight moves to reach a certain square. In the main
index.js file the graph instance that represents a chessboard is made
with edges representing the possible knight movements and the Knight
Travails function uses this to give the desired functionality.